refactor: 공통 기능 모듈화 및 중복 함수 제거
